+++ gravatar/src/main/java/com/gravatar/di/container/GravatarSdkContainer.kt
@@ -2,8 +2,8 @@ package com.gravatar.di.container
 
 import com.gravatar.GravatarConstants.GRAVATAR_API_BASE_URL_V3
 import com.gravatar.moshiadapers.URIJsonAdapter
-import com.gravatar.services.AuthenticationInterceptor
-import com.gravatar.services.AvatarUploadTimeoutInterceptor
+import com.gravatar.services.interceptors.AuthenticationInterceptor
+import com.gravatar.services.interceptors.AvatarUploadTimeoutInterceptor
 import com.gravatar.services.GravatarApi
 import com.squareup.moshi.Moshi
 import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
