+++ gravatar/src/main/java/com/gravatar/services/interceptors/AuthenticationInterceptor.kt
@@ -1,9 +1,9 @@
-package com.gravatar.services
+package com.gravatar.services.interceptors
 
+import com.gravatar.di.container.GravatarSdkContainer
 import okhttp3.Interceptor
 import okhttp3.Request
 import okhttp3.Response
-import com.gravatar.di.container.GravatarSdkContainer.Companion.instance as GravatarSdkDI
 
 internal class AuthenticationInterceptor(private val oauthToken: String? = null) : Interceptor {
     override fun intercept(chain: Interceptor.Chain): Response {
@@ -15,7 +15,7 @@ internal class AuthenticationInterceptor(private val oauthToken: String? = null)
     }
 
     private fun Request.Builder.addAuthorizationHeaderIfPresent(): Request.Builder {
-        return oauthToken?.let { addHeader(it) } ?: GravatarSdkDI.apiKey?.let { addHeader(it) } ?: this
+        return oauthToken?.let { addHeader(it) } ?: GravatarSdkContainer.instance.apiKey?.let { addHeader(it) } ?: this
     }
 
     private fun Request.Builder.addHeader(bearer: String) = addHeader("Authorization", "Bearer $bearer")
